added default admin user, user requires activation, admin user management panel
