+++ docker-compose.prod.yml
@@ -12,6 +12,7 @@ services:
       - PORT=3000
       - NODE_ENV=production
       - CHOKIDAR_POLLING_INTERVAL=60000 
+      - APP_NAME=Discereon
     ports:
       - "3000:3000"
     restart: always
