+++ frontend/server/api/users/[id].js
@@ -20,7 +20,7 @@ export default defineEventHandler(async (event) => {
       // Get user with auth type information (has_password, has_pin)
       const db = getDb();
       const user = db.prepare(`
-        SELECT id, name, avatar, use_auth, isAdmin,
+        SELECT id, name, avatar, use_auth, isAdmin, is_active,
         CASE WHEN password IS NOT NULL AND password != '' THEN 1 ELSE 0 END as has_password,
         CASE WHEN pin IS NOT NULL AND pin != '' THEN 1 ELSE 0 END as has_pin
         FROM users WHERE id = ?
